\name{Bat_distM}
\alias{Bat_distM}
\docType{data}
\title{
  Species pairwise distance matrix for bat species
}
\description{
  This dataset includes a bat species pairwise symmetric distance matrix for 17 species listed in the \code{Bat_incidence_data} dataset. Each element in the matrix represents a Gower distance between two species computed from species traits. \cr
}
\usage{data("Bat_distM")}
\format{
  \code{Bat_distM} is a 17 x 17 (species by species) symmetric data.frame. Each element of the data.frame is between zero (for species pairs with identical traits) and one. \cr
}

\keyword{datasets}
