\name{Bat_incidence_data}
\alias{Bat_incidence_data}
\docType{data}
\title{
  Bat incidence raw data (datatype = "incidence_raw" for TD, PD, and FD)
}
\description{
  This dataset includes bat species incidence data in six sites (B04, B05, B06, B07, H09, P08) with two treatments (Control and Enhanced). In each combinations of six sites and two treatments, there are 4 sampling units.
}
\usage{data("Bat_incidence_data")}
\format{
  \code{Bat_incidence_data} is a matrix where:
  \itemize{
    \item The first three columns represent Site, Treatment, and Patch.
    \item The remaining 17 columns correspond to species names.
  }
}


\keyword{datasets}
