\name{Bat_tree}
\alias{Bat_tree}
\docType{data}
\title{
  Phylogenetic tree for bat incidence data
}
\description{
  This dataset includes a phylogenetic tree spanned by 17 species listed in the \code{Bat_incidence_data} dataset. \cr
}
\usage{data("Bat_tree")}
\format{
  \code{Bat_tree} is a list (phylo tree) with the following phylogenetic information: \cr
  A list with 5 elements: \cr
  $ edge       : int [1:30, 1:2] 18 19 20 20 21 22 23 23 22 21 19 ... \cr
  $ edge.length: num [1:30] 11.686  1.703  8.346  1.346  0.124  1.286  5.589  5.589  6.875 ... \cr
  $ Nnode      : int 16 \cr
  $ tip.label  : chr [1:16] "Myotis.dasycneme" "Myotis.myotis" "Myotis.nattereri" ... \cr
}

\keyword{datasets}
