% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R
\name{DataInfobeta3Dmeta}
\alias{DataInfobeta3Dmeta}
\title{Data information for reference samples}
\usage{
DataInfobeta3Dmeta(
  data,
  diversity,
  datatype,
  PDtree = NULL,
  PDreftime = NULL,
  FDdistM = NULL,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a data.frame. The data frame has study/site and treatment as the first two columns, followed by columns for species names. Here an assemblage refers to a combination of study/site and treatment. \cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a data.frame. The data frame has study/site, treatment, patch as the first three columns, followed by columns for species names.}

\item{diversity}{Selection of diversity type: 'TD' = Taxonomic diversity, 'PD' = Phylogenetic diversity, and 'FD' = Functional diversity.}

\item{datatype}{Data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence data (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection).}

\item{PDtree}{(required only when \code{diversity = "PD"}), a phylogenetic tree in Newick format for all observed species in the pooled data.}

\item{PDreftime}{(required only when \code{diversity = "PD"}), a numerical value specifying reference times for PD. Default is NULL (i.e., the age of the root of PDtree).}

\item{FDdistM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species in the pooled data.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under a specified threshold value, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{FDtype = "AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_value"}), a numerical value between 0 and 1 specifying the tau value (threshold level) that will be used to compute FD. If \code{FDtau = "NULL"} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled data (i.e., quadratic entropy).}
}
\value{
A data.frame including basic data information. \cr
For abundance data, basic information shared by TD, mean-PD and FD includes study/site name
(\code{Site}), treatment name (\code{Treatment}), pooled/joint assemblage (\code{Assemblage}), sample size (\code{n}), observed species
richness (\code{S.obs}), sample coverage estimates of the reference sample (\code{SC(n)}), sample coverage estimate
for twice the reference sample size (\code{SC(2n)}). Other additional information is given below. \cr

\itemize{
(1) TD: the first five species abundance frequency counts in the reference sample (\code{f1–f5}).

(2) Mean-PD: the the observed total branch length in the phylogenetic tree (\code{PD.obs}), the number of
singletons (\code{f1*}) and doubletons (\code{f2*}) in the node/branch abundance set, as well as the total branch
length of those singletons (\code{g1}) and of those doubletons (\code{g2}), and the reference time (\code{Reftime}).

(3) FD (\code{FDtype = "AUC"}): the minimum distance (\code{dmin}) and the maximum distance (\code{dmax}) among
all non-diagonal elements in the distance matrix, and the mean distance between any two individuals
randomly selected from the dataset (\code{dmean}).

(4) FD (\code{FDtype = "tau_value"}): the number of singletons (\code{a1*}) and of doubletons (\code{a2*}) among
the functionally indistinct set at the specified threshold level \code{'Tau'}, as well as the total contribution
of singletons (\code{h1}) and of doubletons (\code{h2}) at the specified threshold level \code{'Tau'}.
}

For incidence data, the basic information for TD includes study/site name (\code{Site}), treatment name (\code{Treatment}),
pooled/joint assemblage (\code{Assemblage}), number of sampling units (\code{T}), total number of incidences (\code{U}), observed
species richness (\code{S.obs}), sample coverage estimates of the reference sample (\code{SC(T)}), sample coverage
estimate for twice the reference sample size (\code{SC(2T)}), as well as the first five species incidence
frequency counts (\code{Q1–Q5}) in the reference sample. For mean-PD and FD, output is similar to that
for abundance data.
}
\description{
\code{DataInfobeta3Dmeta} provides basic data information in each combination of site and treatment for (1) the reference sample in each assemblage, (2) the gamma reference sample in the pooled assemblage, and (3) the alpha reference sample in the joint assemblage.
}
\examples{
## (Data Information) Taxonomic diversity for abundance data
data("Spider_abundance_data")
info_TD_abu <- DataInfobeta3Dmeta(data = Spider_abundance_data, diversity = 'TD', datatype = 'abundance')
info_TD_abu


## (Data Information) Taxonomic diversity for incidence data
data("Bat_incidence_data")
info_TD_inc <- DataInfobeta3Dmeta(data = Bat_incidence_data, diversity = 'TD', datatype = 'incidence_raw')
info_TD_inc


## (Data Information) Mean phylogenetic diversity for abundance data
data("Spider_abundance_data")
data("Spider_tree")
info_PD_abu <- DataInfobeta3Dmeta(data = Spider_abundance_data, diversity = 'PD', datatype = 'abundance',
                                  PDtree = Spider_tree, PDreftime = NULL)
info_PD_abu


## (Data Information) Mean phylogenetic diversity for incidence data
data("Bat_incidence_data")
data("Bat_tree")
info_PD_inc <- DataInfobeta3Dmeta(data = Bat_incidence_data, diversity = 'PD', datatype = 'incidence_raw',
                                  PDtree = Bat_tree, PDreftime = NULL)
info_PD_inc


## (Data Information) Functional diversity for abundance data under a specified threshold level
data("Spider_abundance_data")
data("Spider_distM")
info_FDtau_abu <- DataInfobeta3Dmeta(data = Spider_abundance_data, diversity = 'FD', datatype = 'abundance',
                                     FDdistM = Spider_distM, FDtype = "tau_value", FDtau = NULL)
info_FDtau_abu


## (Data Information) Functional diversity for abundance data when all threshold levels from 0 to 1 are
considered
data("Spider_abundance_data")
data("Spider_distM")
info_FDAUC_abu <- DataInfobeta3Dmeta(data = Spider_abundance_data, diversity = 'FD', datatype = 'abundance',
                                     FDdistM = Spider_distM, FDtype = "AUC", FDtau = NULL)
info_FDAUC_abu


## (Data Information) Functional diversity for incidence data under a specified threshold level
data("Bat_incidence_data")
data("Bat_distM")
info_FDtau_inc <- DataInfobeta3Dmeta(data = Bat_incidence_data, diversity = 'FD', datatype = 'incidence_raw',
                                     FDdistM = Bat_distM, FDtype = "tau_value", FDtau = NULL)
info_FDtau_inc


## (Data Information) Functional diversity for incidence data when all threshold levels from 0 to 1 are
considered
data("Bat_incidence_data")
data("Bat_distM")
info_FDAUC_inc <- DataInfobeta3Dmeta(data = Bat_incidence_data, diversity = 'FD', datatype = 'incidence_raw',
                                     FDdistM = Bat_distM, FDtype = "AUC", FDtau = NULL)
info_FDAUC_inc

}
