\name{iNEXT.meta-package}
\alias{iNEXT.meta-package}
\alias{iNEXT.meta}
\docType{package}
\title{
  Meta analysis for the difference of iNterpolation and EXTrapolation with diversity for three dimensions with two treatments
}
\description{
  \code{iNEXT.meta} (Meta analysis for the difference of iNterpolation and EXTrapolation with diversity for three dimensions with two treatments) is an R package that extends the concepts of \code{iNEXT.3D} (Chao et al., 2021) to meta analysis (fixed effect model). \cr

  \code{iNEXT.meta} is primarily designed to combine the results about difference of diversity with two treatments from two or more different studies. And the package can calculate three dimensions (3D) of diversity includes taxonomic diversity (TD), phylogenetic diversity (PD) and functional diversity (FD) at the particular value of sample coverage or sanple size. \cr


  This package contains two main functions: \cr

  1. \code{iNEXTmeta} estimates the difference of standardized 3D diversity (taxonomic, phylogenetic and functional diversity with common sample size or sample coverage), and fit the fixed effect model to perform meta analysis. \cr

  2. \code{ggiNEXTmeta} visualizes the output from the function \code{iNEXTmeta} by providing forest plot of the difference of standardized 3D diversity with two treatments in each study/site and meta analysis (fixed effect model). \cr

}

\keyword{ package }
