% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{iNEXT.meta}
\alias{iNEXT.meta}
\title{iNEXT.meta() is a function that estimates the difference of standardized 3D (taxonomic, phylogenetic and
functional) gamma, alpha, beta diversity, and four dissimilarity measures with two treatments (e.g., enhanced vs. control), and perform meta analysis for several
studies/sites.}
\usage{
iNEXT.meta(
  data,
  diversity = "TD",
  order.q = 0,
  datatype = "incidence_raw",
  level = NULL,
  nboot = 10,
  treatment_order,
  conf = 0.95,
  PDtree,
  PDreftime = NULL,
  PDtype = "meanPD",
  FDdistM,
  FDtype = "AUC",
  FDtau = NULL,
  FDcut_number = 30
)
}
\arguments{
\item{data}{(a) For datatype = "abundance", data can be input as a matrix/data.frame (species by assemblages).
Here an assemblage refers to a combination of site and treatment.
The first row lists the names of sites and second row lists the names of treatment. 
(b) For datatype = "incidence_raw", data can be input as a list with several lists (assemblages) of
matrices/data.frames, each matrix represents species-by-sampling units incidence data . 
The first list must be a dataframe with names of site-by-treatment for each list.}

\item{diversity}{selection of diversity type: 'TD' = Taxonomic diversity, 'PD' = Phylogenetic diversity, 
and 'FD' = Functional diversity.}

\item{order.q}{a number specifying the diversity order. Default is \code{order.q = 0}.}

\item{datatype}{data type of input data: individual-based abundance data (datatype = "abundance"), 
species by sampling-units incidence matrix (datatype = "incidence_raw") with all entries being 0 (non-detection)
or 1 (detection).}

\item{level}{a number specifying a particular value of sample coverage (between 0 and 1) that will be used to compute standardized 3D estimates. 
If level = NULL, then this function computes the standardized 3D diversity estimates and four dissimilarity measures
for the minimal quartile among all SC(2n) values that are greater than 0.5.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. 
Bootstrap replications are generally time consuming. 
Default is nboot = 10. If more accurate results are required, set nboot = 100 (or nboot = 200).}

\item{treatment_order}{a character vector for the names of treatment. The difference of standardized 3D 
diversity and the four dissimilarity measures will be computed as diversity of the first treatment minus the diversity of second treatment.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{PDtree}{(required only when diversity = "PD"), a phylogenetic tree in Newick format for all observed
species in the pooled data.}

\item{PDreftime}{(required only when diversity = "PD"), a number specifying reference 
times for PD. Default is NULL (i.e., the age of the root of PDtree).}

\item{PDtype}{(required only when diversity = "PD"), select PD type: PDtype = "PD" (effective total branch length)
or PDtype = "meanPD" (effective number of equally divergent lineages). Default is "meanPD", where meanPD = PD/tree depth.}

\item{FDdistM}{(required only when diversity = "FD"), a species pairwise distance matrix for all species in
the pooled data.}

\item{FDtype}{(required only when diversity = "FD"), select FD type: FDtype = "tau_value" for FD under 
specified a threshold value, or FDtype = "AUC" (area under the curve of tau-profile) for an overall FD which
integrates all threshold values between zero and one. Default is "AUC".}

\item{FDtau}{(required only when diversity = "FD" and FDtype = "tau_value"), a number between 0 and 1 
specifying tau value (threshold levels). If NULL (default), then threshold is set to be the mean distance 
between any two individuals randomly selected from the pooled data (i.e., quadratic entropy).}

\item{FDcut_number}{(argument only for diversity = "FD" and FDtype = "AUC"), a numeric number to cut [0, 1] interval into equal-spaced sub-intervals to obtain the AUC value by integrating the tau-profile. Equivalently, the number of tau values that will be considered to compute the integrated AUC value. Default is FDcut_number = 30. A larger value can be set to obtain more accurate AUC value.}
}
\value{
a list. The list contains eight dataframes (coverage summary, gamma, alpha, beta diversity, and four dissimilarity measures) that show the results for each site.
}
\description{
iNEXT.meta() is a function that estimates the difference of standardized 3D (taxonomic, phylogenetic and
functional) gamma, alpha, beta diversity, and four dissimilarity measures with two treatments (e.g., enhanced vs. control), and perform meta analysis for several
studies/sites.
}
