\name{Spider_abundance_data}
\alias{Spider_abundance_data}
\docType{data}
\title{
  Spider abundance data (datatype = "abundance" for TD, PD, and FD)
}
\description{
  This dataset includes spider species abundance data in eleven sites (B04, B05, B06, B07, H09, L11, P08, S10, U01, U02, U03) with two treatments (Control and Enhanced).
}
\usage{data("Spider_abundance_data")}
\format{
  \code{Spider_abundance_data} is a matrix where:
  \itemize{
    \item The first two columns represent Site, Treatment.
    \item The remaining 199 columns correspond to species names.
  }
}


\keyword{datasets}
