\name{Spider_tree}
\alias{Spider_tree}
\docType{data}
\title{
  Phylogenetic tree for spider abundance data
}
\description{
  This dataset includes a phylogenetic tree spanned by 199 species listed in the \code{Spider_abundance_data} dataset. \cr
}
\usage{data("Spider_tree")}
\format{
  \code{Spider_tree} is a list (phylo tree) with the following phylogenetic information: \cr
  A list of 5: \cr
  $ edge       : int [1:377, 1:2] 200 201 202 203 204 205 206 206 206 205 207 ... \cr
  $ edge.length: num [1:30] 0.166  0.083  0.083  0.083  0.250  0.195  0.055  0.055  0.139 ... \cr
  $ Nnode      : int 179 \cr
  $ node.label : 100 97 100.2 100.3 100.4 9 100.5 36 58 100.6 ... \cr
  $ tip.label  : chr [1:16] "Araniella_opisthographa" "Araniella_alpica" "Zilla_diodia" ... \cr
}

\keyword{datasets}
