\name{bat_incidence_data}
\alias{bat_incidence_data}
\docType{data}
\title{
  Bat incidence raw data (datatype = "incidence_raw" for TD, PD, and FD)
}
\description{
  This dataset includes bat species incidence data in eight sites (B04, B05, B06, B07, P08, U01, U02, U03) with two treatments (Control and Enhanced). In each combinations of eight sites and two treatment, there is about 33 to 45 sampling units.
}
\usage{data("bat_incidence_data")}
\format{
  \code{bat_incidence_data} is a list with 16 species-by-sampling units data.frames/matrices. \cr
  A list of 16 matrices: \cr
  $ Site_B04_Enhanced is a matrix with 16 species (rows) and 36 columns (sampling units). \cr
  $ Site_B05_Enhanced is a matrix with 16 species (rows) and 33 columns (sampling units). \cr
  $ Site_B06_Enhanced is a matrix with 16 species (rows) and 34 columns (sampling units). \cr
  $ Site_B07_Enhanced is a matrix with 16 species (rows) and 35 columns (sampling units). \cr
  $ Site_P08_Enhanced is a matrix with 16 species (rows) and 34 columns (sampling units). \cr
  $ Site_U01_Enhanced is a matrix with 16 species (rows) and 45 columns (sampling units). \cr
  $ Site_U02_Enhanced is a matrix with 16 species (rows) and 40 columns (sampling units). \cr
  $ Site_U03_Enhanced is a matrix with 16 species (rows) and 43 columns (sampling units). \cr
  $ Site_B04_Control is a matrix with 16 species (rows) and 34 columns (sampling units). \cr
  $ Site_B05_Control is a matrix with 16 species (rows) and 36 columns (sampling units). \cr
  $ Site_B06_Control is a matrix with 16 species (rows) and 36 columns (sampling units). \cr
  $ Site_B07_Control is a matrix with 16 species (rows) and 35 columns (sampling units). \cr
  $ Site_P08_Control is a matrix with 16 species (rows) and 36 columns (sampling units). \cr
  $ Site_U01_Control is a matrix with 16 species (rows) and 44 columns (sampling units). \cr
  $ Site_U02_Control is a matrix with 16 species (rows) and 37 columns (sampling units). \cr
  $ Site_U03_Control is a matrix with 16 species (rows) and 43 columns (sampling units). \cr
}

\keyword{datasets}
