\name{bat_tree}
\alias{bat_tree}
\docType{data}
\title{
  Phylogenetic tree for bat incidence data
}
\description{
  This dataset includes a phylogenetic tree spanned by 16 species listed in the \code{bat_incidence_data} dataset.  \cr
}
\usage{data("bat_tree")}
\format{
  \code{bat_tree} is a list (phylo tree) with the following phylogenetic information: \cr
  A list of 5: \cr
  $ edge       : int [1:30, 1:2] 17 18 19 19 20 21 18 19 1 20 21 ... \cr
  $ edge.length: num [1:30] 11.686  1.703  8.346  1.346  0.124  1.286  5.589  5.589  6.875 ... \cr
  $ Nnode      : int 15 \cr
  $ tip.label  : chr [1:16] "Myotis.dasycneme" "Myotis.myotis" "Myotis.nattereri" ... \cr
}

\keyword{datasets}
