% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{ggiNEXTmeta}
\alias{ggiNEXTmeta}
\title{ggiNEXTmeta() is a function that provides forest plot for the difference of standardized 3D
diversity and four dissimilarity measures with two treatments.}
\usage{
ggiNEXTmeta(data, range, num_round, type = NULL)
}
\arguments{
\item{data}{the outcome of the iNEXT.meta function.}

\item{range}{the range of the forest plot.}

\item{num_round}{a number that the values show on the plot are rounded to the specified number of decimal places.}

\item{type}{(argument only for the output from iNEXT.meta), selection of diversity type for plotting forest plot: 'Gamma' = Gamma diversity, 'Alpha' = Alpha diversity, 'Beta' = Beta diversity,
1-C' = Sørensen-type non-overlap dissimilarity, '1-U' = Jaccard-type non-overlap dissimilarity, '1-V' = Sørensen-type turnover dissimilarity, and '1-S' = Jaccard-type turnover dissimilarity.}
}
\value{
A forest plot that visualizing the output of iNEXTbeta3Dmeta. In the plot, it shows the difference of diversity between two treatments for each study/site and meta analysis (fixed-effects model).
}
\description{
ggiNEXTmeta() is a function that provides forest plot for the difference of standardized 3D
diversity and four dissimilarity measures with two treatments.
}
