% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R
\name{ggiNEXTmeta}
\alias{ggiNEXTmeta}
\title{forest plot for the difference of standardized 3D diversity with two treatments}
\usage{
ggiNEXTmeta(data, range, num_round)
}
\arguments{
\item{data}{the outcome of the iNEXTmeta function.}

\item{range}{the range of the forest plot.}

\item{num_round}{a number that the values show on the plot are rounded to the specified number of decimal places.}
}
\value{
a forest plot that visualizing the output of iNEXTmeta.
}
\description{
\code{ggiNEXTmeta} is a function that provides forest plot for the difference of standardized 3D diversity with two treatments.
}
\examples{
data("bat")
output <- iNEXTmeta(bat$data, diversity="TD", order.q=0,
                    datatype="incidence_raw", base="coverage", level=0.9,
                    treatment_order=c("Enhanced","Control"))
ggiNEXTmeta(output, c(-11,16), 3)
}
